requirements.txt
arrow==1.2.3
astroid==2.12.10
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
dill==0.3.5.1
Flask==2.1.3
h11==0.13.0
html5validator==0.4.2
idna==3.4
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
outcome==1.2.0
packaging==21.3
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.9.1
pydocstyle==6.1.1
pylint==2.15.3
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.3
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tomli==2.0.1
tomlkit==0.11.4
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
Werkzeug==2.0.3
wrapt==1.14.1
wsproto==1.2.0
Flask-OAuthlib
Flask-Login
Flask-Session
PyPDF2
nltk